refactor away index.ts due to circular deps issues
