+++ src/app.module.ts
@@ -3,7 +3,7 @@ import { ConfigModule } from '@nestjs/config'
 import { ScheduleModule } from '@nestjs/schedule'
 
 import { ApiModule } from '@src/module.api'
-import { DatabaseModule } from '@src/module.database'
+import { DatabaseModule } from '@src/module.database/module'
 import { DeFiDModule } from '@src/module.defid'
 import { HealthModule } from '@src/module.health'
 import { AppConfiguration } from '@src/app.configuration'
