+++ src/module.database/module.ts
@@ -4,9 +4,9 @@ import { ConfigService } from '@nestjs/config'
 import { FactoryProvider } from '@nestjs/common/interfaces/modules/provider.interface'
 import { Database } from '@src/module.database/database'
 import { LevelDatabase } from '@src/module.database/provider.level/level.database'
-import { LevelDatabaseModule } from '@src/module.database/provider.level'
+import { LevelDatabaseModule } from '@src/module.database/provider.level/module'
 import { MemoryDatabase } from '@src/module.database/provider.memory/memory.database'
-import { MemoryDatabaseModule } from '@src/module.database/provider.memory'
+import { MemoryDatabaseModule } from '@src/module.database/provider.memory/module'
 
 /**
  * Runtime DatabaseProvider, dynamically inject a database type based on config in the database.provider.
